test(testkit): move cross-crate scenarios to integration tests
