+++ crates/reinhardt-testkit/src/auth/server_fn_builder.rs
@@ -70,55 +70,3 @@ impl ServerFnAuthBuilder {
 		self.ctx
 	}
 }
-
-#[cfg(test)]
-mod tests {
-	use std::sync::Arc;
-
-	use reinhardt_di::SingletonScope;
-	use serde_json::json;
-	use uuid::Uuid;
-
-	use super::*;
-	use crate::auth::JwtTestConfig;
-
-	struct FixtureUser(Uuid);
-
-	impl ForceLoginUser for FixtureUser {
-		fn session_user_id(&self) -> String {
-			self.0.to_string()
-		}
-	}
-
-	#[test]
-	fn server_fn_auth_builder_preserves_session_and_jwt_identity_overrides() {
-		// Arrange
-		let user = FixtureUser(Uuid::from_u128(0x8a5d));
-		let scope = Arc::new(SingletonScope::new());
-
-		// Act
-		let session_env = ServerFnTestContext::new(scope.clone())
-			.auth()
-			.session(&user)
-			.with_staff(true)
-			.with_superuser(true)
-			.done()
-			.build();
-		let jwt_env = ServerFnTestContext::new(scope)
-			.auth()
-			.jwt(&user, &JwtTestConfig::default())
-			.with_staff(true)
-			.with_superuser(true)
-			.done()
-			.build();
-
-		// Assert
-		for env in [session_env, jwt_env] {
-			let session = env.mock_session.unwrap();
-			assert_eq!(session.user.as_ref().unwrap().id, user.0);
-			assert_eq!(session.get_raw("user_id"), Some(&json!(user.0.to_string())));
-			assert_eq!(session.get_raw("is_staff"), Some(&json!(true)));
-			assert_eq!(session.get_raw("is_superuser"), Some(&json!(true)));
-		}
-	}
-}
