+++ crates/reinhardt-testkit/src/messages.rs
@@ -435,51 +435,4 @@ mod tests {
 		assert!(!filtered.contains("testing::"));
 		assert!(!filtered.contains("unittest::"));
 	}
-
-	#[test]
-	fn message_mixin_delegates_tag_matching_and_stack_trace_modes() {
-		// Arrange
-		let mut saved = Message::new(Level::Success, "Profile saved".to_string());
-		saved.extra_tags = vec!["profile".to_string(), "toast".to_string()];
-		let messages = vec![saved];
-		let expected_messages = vec![(Level::Success, "Profile saved".to_string())];
-		let expected_tags = vec![(
-			Level::Success,
-			"Profile saved".to_string(),
-			vec!["profile".to_string(), "toast".to_string()],
-		)];
-		let preserving = MessagesTestMixin::new();
-		let filtering = MessagesTestMixin::with_settings(true);
-		let stack_trace = "at app::handler\nat suite::assert_case\nat testing::helper\nat main";
-
-		// Act
-		let exists = preserving.assert_message_exists(&messages, Level::Success, "Profile saved");
-		let ordered = preserving.assert_messages(&messages, &expected_messages, true);
-		let tagged = preserving.assert_messages_with_tags(&messages, &expected_tags);
-		let missing = preserving.assert_messages_with_tags(
-			&messages,
-			&[(
-				Level::Success,
-				"Profile saved".to_string(),
-				vec!["other".to_string()],
-			)],
-		);
-		let preserved = preserving.filter_stack_trace(stack_trace);
-		let filtered = filtering.filter_stack_trace(stack_trace);
-
-		// Assert
-		exists.expect("matching message should be found");
-		ordered.expect("ordered message should match");
-		tagged.expect("matching message tags should be found");
-		match missing {
-			Err(MessageAssertionError::MessageNotFound { message }) => assert_eq!(
-				message,
-				"Message with level Success, text 'Profile saved', and tags [\"other\"] not found"
-			),
-			Err(other) => panic!("unexpected message assertion error: {other}"),
-			Ok(()) => panic!("expected mismatched tags to be rejected"),
-		}
-		assert_eq!(preserved, stack_trace);
-		assert_eq!(filtered, "at app::handler\nat main");
-	}
 }
