+++ tests/integration/tests/testkit.rs
@@ -6,6 +6,9 @@ mod auth_integration;
 #[path = "testkit/client_integration.rs"]
 mod client_integration;
 
+#[path = "testkit/messages_integration.rs"]
+mod messages_integration;
+
 #[path = "testkit/static_files_integration.rs"]
 mod static_files_integration;
